Add drawdown to `BasePrice` guis
Adds option to `BasePrice` for subclasses to include a drawdown
plot and use a `ToggleButtons` widget to select the period over
which to evaluate the high.

Implementation includes defining a new `prices` attribute on
`guis.BasePrice` to hold the prices as currently represented on
the chart (the structure of this data can differ from that actually
passed through to the chart, this can be an issue as evaluating
the drawdown requires knowledge of the 'high' prices whilst
some charts, for example `charts.Line`  requires only the close
price. The `prices` attribure provides for retaining knowledge of
all the price data from which new drawdown evaluations can be
made).

Rearranges order of parameters on some internal 'update'
methods on charts.py and guis.py.

Also refactors base of `IntervalSelector` to a dedicated
`ToggleButtonsHandled` class which the new
`DrawdownSelector` also uses.